+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 1.0.1 - TBD
+
+Initial release.
+
+### Added
+
+- Nothing.
+
+### Changed
+
+- Apply patches before build of configuration files.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- Listener sample for overriding of default `config.php` configuration file.
+- Useless import of the `File::Temp` package (package handler).
+- CS fixes
+
 ## 1.0.0 - 20190403
 
 Initial release.